2023/test_day16.py
"""Tests Day 16."""
import pytest
from puzzle_data import PuzzleData
from day16 import Day16


# pylint: disable=invalid-name,missing-function-docstring


# The below will change dir to the directory of this file so pytest
# can be run from any other directory.
@pytest.fixture(autouse=True)
def change_test_dir(request, monkeypatch):
    monkeypatch.chdir(request.fspath.dirname)


@pytest.fixture(name="puzzle")
def fixture_puzzle():
    return Day16(PuzzleData("data/sample/16"))


energizedTopLeftE = """\
######....
.#...#....
.#...#####
.#...##...
.#...##...
.#...##...
.#..####..
########..
.#######..
.#...#.#..
"""


def test_energizedTopLeftE(puzzle):
    puzzle.lc.project(puzzle.topLeftE)
    assert puzzle.lc.energized_viz() == energizedTopLeftE


def test_part1_solution(puzzle):
    puzzle.calculate_part('a')
    assert puzzle.results['a'] == puzzle.puzzle_data.solution('a')


def test_part2_solution(puzzle):
    puzzle.calculate_part('b')
    assert puzzle.results['b'] == puzzle.puzzle_data.solution('b')
